+++ harmony_browse_image_generator/exceptions.py
@@ -5,13 +5,19 @@
 from harmony.util import HarmonyException
 
 
-class HyBIGException(HarmonyException):
+class HyBIGError(HarmonyException):
     """Base service exception."""
     def __init__(self, message=None):
         super().__init__(message, 'sds/harmony-browse-image-generator')
 
 
-class HyBIGInvalidMessage(HarmonyException):
+class HyBIGNoColorInformation(HarmonyException):
+    """Used to describe missing color information."""
+    def __init__(self, message=None):
+        super().__init__(message, 'sds/harmony-browse-image-generator')
+
+
+class HyBIGInvalidMessageError(HarmonyException):
     """Input Harmony Message could not be used as presented."""
     def __init__(self, message=None):
         super().__init__(message, 'sds/harmony-browse-image-generator')